requirements.txt
Flask==3.0.1
Jinja2==3.1.3
numpy==1.26.3
pandas==2.2.0
pip==23.2.1
pyarrow==15.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
tzdata==2023.4
Werkzeug==3.0.1
gunicorn
snowflake-connector-python